+++ Event/DataTablesEvents.php
@@ -16,6 +16,7 @@ namespace WBW\Bundle\JQuery\DataTablesBundle\Event;
  *
  * @author webeweb <https://github.com/webeweb/>
  * @package WBW\Bundle\JQuery\DataTablesBundle\Event
+ * @deprecated since 3.4.0 use "WBW\Bundle\JQuery\DataTablesBundle\WBWJQueryDataTablesEvents" instead.
  */
 class DataTablesEvents {
 
@@ -24,62 +25,62 @@ class DataTablesEvents {
      *
      * @var string
      */
-    const DATATABLES_PRE_DELETE = "wbw.jquery.datatables.event.pre_delete";
+    const DATATABLES_PRE_DELETE = WBWJQueryDataTablesEvents::DATATABLES_PRE_DELETE;
 
     /**
      * DataTables "pre edit".
      *
      * @var string
      */
-    const DATATABLES_PRE_EDIT = "wbw.jquery.datatables.event.pre_edit";
+    const DATATABLES_PRE_EDIT = WBWJQueryDataTablesEvents::DATATABLES_PRE_EDIT;
 
     /**
      * DataTables "pre export".
      *
      * @var string
      */
-    const DATATABLES_PRE_EXPORT = "wbw.jquery.datatables.event.pre_export";
+    const DATATABLES_PRE_EXPORT = WBWJQueryDataTablesEvents::DATATABLES_PRE_EXPORT;
 
     /**
      * DataTables "pre index".
      *
      * @var string
      */
-    const DATATABLES_PRE_INDEX = "wbw.jquery.datatables.event.pre_index";
+    const DATATABLES_PRE_INDEX = WBWJQueryDataTablesEvents::DATATABLES_PRE_INDEX;
 
     /**
      * DataTables "pre show".
      *
      * @var string
      */
-    const DATATABLES_PRE_SHOW = "wbw.jquery.datatables.event.pre_show";
+    const DATATABLES_PRE_SHOW = WBWJQueryDataTablesEvents::DATATABLES_PRE_SHOW;
 
     /**
      * DataTables "post delete".
      *
      * @var string
      */
-    const DATATABLES_POST_DELETE = "wbw.jquery.datatables.event.post_delete";
+    const DATATABLES_POST_DELETE = WBWJQueryDataTablesEvents::DATATABLES_POST_DELETE;
 
     /**
      * DataTables "post edit".
      *
      * @var string
      */
-    const DATATABLES_POST_EDIT = "wbw.jquery.datatables.event.post_edit";
+    const DATATABLES_POST_EDIT = WBWJQueryDataTablesEvents::DATATABLES_POST_EDIT;
 
     /**
      * DataTables "post export".
      *
      * @var string
      */
-    const DATATABLES_POST_EXPORT = "wbw.jquery.datatables.event.post_export";
+    const DATATABLES_POST_EXPORT = WBWJQueryDataTablesEvents::DATATABLES_POST_EXPORT;
 
     /**
      * DataTables "post show".
      *
      * @var string
      */
-    const DATATABLES_POST_INDEX = "wbw.jquery.datatables.event.post_index";
+    const DATATABLES_POST_INDEX = WBWJQueryDataTablesEvents::DATATABLES_POST_INDEX;
 
 }